Provide a functioning `getAuthenticatedRequest()` for Instagram
TL;DR: This adds the `access_token` query string parameter to the URL in order
to return a properly authenticated Instagram request. More background
details follow.

AbstractProvider provides a basic `getAuthenticatedRequest()` method,
which the Instagram provider was using unchanged. This method creates a
request, and in doing so, merges in the headers array returned by
`getAuthorizationHeaders()`, which is an empty array in the
AbstractProvider.

The base package provides a BearerAuthorizationTrait that providers may
use in order to override `getAuthorizationheaders()` and set an
Authorization header using Bearer token authentication on all requests,
in order to request a properly authenticated request.

Instagram, however, does not support the use of the Authorization header
for authenticated requests. Instead, the `access_token` parameter must
be present in the query string of all authenticated requests.